chore: Removed dummy Bank POJO and replaced with the ../models/Bank.java
