+++ app/src/main/java/co/sigmoidlabs/bankussdtoolbox/data/model/Bank.java
@@ -1,5 +1,7 @@
 package co.sigmoidlabs.bankussdtoolbox.data.model;
 
+import android.os.Parcel;
+import android.os.Parcelable;
 import android.support.annotation.ColorInt;
 
 import java.util.List;
@@ -8,7 +10,7 @@ import java.util.List;
  * Created by Efe on 04/08/2016.
  */
 
-public class Bank {
+public class Bank implements Parcelable{
 
     String key;
     String name;
@@ -16,4 +18,36 @@ public class Bank {
     @ColorInt int accentColor;
 
     List<Action> actions;
+
+    protected Bank(Parcel in) {
+        key = in.readString();
+        name = in.readString();
+        color = in.readInt();
+        accentColor = in.readInt();
+    }
+
+    public static final Creator<Bank> CREATOR = new Creator<Bank>() {
+        @Override
+        public Bank createFromParcel(Parcel in) {
+            return new Bank(in);
+        }
+
+        @Override
+        public Bank[] newArray(int size) {
+            return new Bank[size];
+        }
+    };
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(Parcel parcel, int i) {
+        parcel.writeString(key);
+        parcel.writeString(name);
+        parcel.writeInt(color);
+        parcel.writeInt(accentColor);
+    }
 }
