+++ app/src/main/java/co/sigmoidlabs/bankussdtoolbox/widget/RoundedLinearLayout.java
@@ -42,6 +42,11 @@ public class RoundedLinearLayout extends ForegroundLinearLayout {
     }
 
     @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        super.onMeasure(widthMeasureSpec, widthMeasureSpec);
+    }
+
+    @Override
     protected void onSizeChanged(int w, int h, int oldw, int oldh) {
         super.onSizeChanged(w, h, oldw, oldh);
